Update video.py
Added a few more imports.

Set more variables to be able to have timestamps as file names.

Learned how to join paths and file names

Set the rsolution to 1080p60 for testing - had chatgpt help with the framerate.

Changed the runtime to 1 hour.

If you want this to run continuously, call it in crontab so that you get files however often you desire.  Just make sure you adjust the time.sleep(3600) to what you want as well.